pyproject.toml
[project]
name = "ai-meeting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.65.0",
    "notion-client>=3.0.0",
    "numpy<2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "soundcard>=0.4.5",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]
